Main TimeCycle Function

Main function of the TimeCycle package used for detecting rhythmic signals in time-series gene expression sets. For additional help with parameter selection, see TimeCycle's vignette: vignette("TimeCycle").

TimeCycle(
  data,
  repLabel,
  resamplings = 10000,
  minLag = 2,
  maxLag = 5,
  period = 24,
  cores = parallel::detectCores() - 2,
  laplacian = T,
  linearTrend = F
)

Arguments

data
a data.frame of numeric gene expression over time (row = genes x col = ZT times).

repLabel
a vector defining the number of replicates at each time points.

resamplings
a numeric specifying the number of resamplings to use in the null-distribution. Default is 10000.

minLag
a numeric specifying the min lag to check in the 3-D embedding. Default is 2.

maxLag
a numeric specifying the max lag to check in the 3-D embedding. Default is 5.

period
a numeric specifying the period of interest in hours for rhythm detection. Default is 24.

cores
a numeric specifying the number of parallel cores to use. Default number of cores is parallel::detectedCores() - 2.

laplacian
a logical scalar. Should the Laplacian Eigenmaps be used for dimensionality reduction? Default TRUE

linearTrend
a logical scalar. Should TimeCycle Prioritize detecting linear trending signals? Default FALSE. Not recommended to change from default FALSE - will increases false positives rate. See vignette("TimeCycle") for more details.

Value

a tidy data.frame of processed results for each gene:

sampleName	perScore	pVals	pVals.adj	Period.in.Hours	Amp	Phase.In.Hours

the gene name	the median persistence score across all lags (min to max)	raw empirical p-value	FDR adjusted p-value	period (h)	amplitude	phase (h)

References

Link to TimeCycle Paper!

TDA Package References

Fasy, Brittany & Kim, Jisu & Lecci, Fabrizio & Maria, Clément. (2014). "Introduction to the R package TDA".

Maria C (2014). "GUDHI, Simplicial Complexes and Persistent Homology Packages." https://project.inria.fr/gudhi/software/

Morozov D (2007). "Dionysus, a C++ library for computing persistent homology". http://www.mrzv.org/software/dionysus/

Edelsbrunner H, Harer J (2010). "Computational topology: an introduction." American Mathematical Society.

Fasy B, Lecci F, Rinaldo A, Wasserman L, Balakrishnan S, Singh A (2013). "Statistical Inference For Persistent Homology." (arXiv:1303.7117). Annals of Statistics.

Examples

# use built in zhang2014 data set sampled every
# 2 hours for 48 hours (i.e. 24 time points with 1 replicate each).
# Search for genes with period of 24 hours.

#set seed for reproducibility with random variables in example usage
set.seed(1234)

TimeCycleResults <- TimeCycle(data = zhang2014[1:100,],
                              repLabel = rep(1,24),
                              period = 24,
                              cores = 2,
                              resamplings = 10)

#> 
#>       ########################################################################################
#>       ###      ████████ ██ ███    ███ ███████  ██████ ██    ██  ██████ ██      ███████     ###
#>       ###         ██    ██ ████  ████ ██      ██       ██  ██  ██      ██      ██          ###
#>       ###         ██    ██ ██ ████ ██ █████   ██        ████   ██      ██      █████       ###
#>       ###         ██    ██ ██  ██  ██ ██      ██         ██    ██      ██      ██          ###
#>       ###         ██    ██ ██      ██ ███████  ██████    ██     ██████ ███████ ███████     ###
#>       ########################################################################################
#> [1] "Starting TimeCycle"
#> [1] "Pre-Processing Data"

#> Warning: replacing previous import ‘vctrs::data_frame’ by ‘tibble::data_frame’ when loading ‘dplyr’

#> Registered S3 method overwritten by 'quantmod':
#>   method            from
#>   as.zoo.data.frame zoo 

#> [1] "Computing Periods"
#> [1] "Pre-Processing Null Distribution"
#> [1] "Computing Null Distribution"
#> [1] "Computing Persistence Scores"
#> [1] "Calculating p-values"
#> [1] "TimeCycle Completed"
#> [1] "Analysis Time: 00:00:01"

# Check number of genes with FDR < 0.05 and period between 22 to 26 hours.
library(tidyverse)

#> ── Attaching packages ─────────────────────────────────────── tidyverse 1.3.0 ──

#> ✓ ggplot2 3.3.2     ✓ purrr   0.3.4
#> ✓ tibble  3.0.4     ✓ dplyr   1.0.1
#> ✓ tidyr   1.1.1     ✓ stringr 1.4.0
#> ✓ readr   1.3.1     ✓ forcats 0.5.0

#> ── Conflicts ────────────────────────────────────────── tidyverse_conflicts() ──
#> x dplyr::filter() masks stats::filter()
#> x dplyr::lag()    masks stats::lag()

TimeCycleResults %>%
   filter(22 < Period.in.Hours & Period.in.Hours < 26) %>%
   filter(pVals.adj < 0.05) %>%
   glimpse()

#> Rows: 0
#> Columns: 7
#> $ sampleNames     <fct> 
#> $ perScore        <dbl> 
#> $ pVals           <dbl> 
#> $ pVals.adj       <dbl> 
#> $ Period.in.Hours <dbl> 
#> $ Amp             <dbl> 
#> $ Phase.in.Hours  <dbl>