Computes Persistence Scores For a Data.Frame of Time-Series Across Multiple Lags

Takes a data.frame of numeric gene expression over time (genes X ZT times) and computes the persistence score using getPersistence. For a given gene, each lag (min to max) is used to transform the expression into a 3-D embedded space via time-delay embedding. A non-linear dimension reduction technique (laplacian eigenmaps) is used to transfrom the 3-D embedding to a 2-D embedding. Finally, the persistence score of the 2-D embedding is calculated via persistence homology. The median persistence score across all lags (min to max) for each gene is returned as a numeric vector. For more details see TimeCycle's vignette: vignette("TimeCycle").

computePersistence(
  data,
  minLag = 2,
  maxLag = 5,
  cores = parallel::detectCores() - 2,
  laplacian = T
)

Arguments

data
a data.frame of numeric gene expression over time (row = genes x col = ZT times).

minLag
a numeric specifying the min lag to check in the 3-D embedding. Default is 2.

maxLag
a numeric specifying the max lag to check in the 3-D embedding. Default is 5.

cores
a numeric specifying the number of parallel cores to use. Default number of cores is parallel::detectedCores() - 2.

laplacian
a logical scalar. Should the Laplacian Eigenmaps be used for dimensionality reduction? Default TRUE.

Value

a vector of the median persistence score across lags (minLag to maxLag) for each gene in data

References

Fasy, Brittany & Kim, Jisu & Lecci, Fabrizio & Maria, Clément. (2014). "Introduction to the R package TDA".

Maria C (2014). "GUDHI, Simplicial Complexes and Persistent Homology Packages." https://project.inria.fr/gudhi/software/ .

Morozov D (2007). "Dionysus, a C++ library for computing persistent homology". http://www.mrzv.org/software/dionysus/

Edelsbrunner H, Harer J (2010). "Computational topology: an introduction." American Mathematical Society.

Fasy B, Lecci F, Rinaldo A, Wasserman L, Balakrishnan S, Singh A (2013). "Statistical Inference For Persistent Homology." (arXiv:1303.7117). Annals of Statistics.

See also